Mini-Compiler/requirements.txt
antlr4-python3-runtime==4.9.2
pytest==7.4.0 